requirements.txt
python-dotenv
fastapi 
uvicorn 
openai 
python-multipart
pydub
pytest-asyncio
pytest 
httpx
